XSD Schema Change
The URL to the v1.3 .xsd GIBS colormap schema was returning a 404, so I had to find a new URL to get the XSD from. This one comes from the GIBS GitHub repo, which will probably remain fairly stable. Another option is to download the .xsd locally and keep in in here.